requirements.txt
playwright
markdown
markdown-include
rich
python-dotenv
requests
Pillow
fpdf2
PyPDF2
